libefiwrapper/bs: make ExitBootService return EXIT_SUCCESS
My original thought was it was up the role of an efiwrapper platform
driver to take care of adding the support for a default EXIT_SUCCESS
behavior for ExitBootService.  But it appears to me now that it is
putting an necessary burden on platforms which do not need much
specific efiwrapper driver.  The default behavior should be to return
EXIT_SUCCESS and if a platform requires a specific behavior it can be
implemented in a specific platform driver.

Change-Id: If8e3f2204e54f50c22d51200f7a04c437094af0a
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-59336
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/621748